backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# AI & ML
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.0
langchain==0.0.340
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
scikit-learn==1.3.2

# Medical NLP & Processing
scispacy==0.5.3
en_core_sci_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_md-0.5.3.tar.gz
medcat==1.7.0
negspacy==1.0.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
python-magic==0.4.27

# Vector Database
pinecone-client==2.2.4
chromadb==0.4.18
faiss-cpu==1.7.4

# Search & Indexing
elasticsearch==8.11.0
whoosh==2.7.4

# Task Queue
celery==5.3.4
flower==2.0.1
redis==5.0.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# Networking & External APIs
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development & Debugging
python-dotenv==1.0.0
ipython==8.17.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0
python-json-logger==2.0.7

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Neurosurgery-Specific Libraries
nibabel==5.1.0  # Neuroimaging
dipy==1.7.0     # Diffusion imaging
nilearn==0.10.2  # Machine learning for neuroimaging
mne==1.5.1      # MEG/EEG processing
pydicom==2.4.3  # DICOM medical imaging

# Graph & Network Analysis (for citation networks)
networkx==3.2.1
pyvis==0.3.2
plotly==5.18.0

# Cache
diskcache==5.6.3
cachetools==5.3.2

# Date & Time
python-dateutil==2.8.2
pytz==2023.3.post1

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0
tabulate==0.9.0